Have the DummyLanguage define an xpath fn
 - Make it readily available for all tests instead of defined outside of
   it only for Saxon tests.
